Add files via upload
COVID-19, a novel coronavirus, is currently a major worldwide threat. It has infected more than a million people globally leading to hundred-thousands of deaths. In such grave circumstances, it is very important to predict the future death numbers to support prevention of the disease and aid in the healthcare service preparation. I've used the data from kaggle and by fitting ARIMA Model, I've predicted the death rates in India.